pyproject.toml
[project]
name = "llama3-dpo-alignment-pipeline"
version = "0.1.0"
description = "A professional pipeline for Llama-3 DPO alignment using Unsloth"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch",
    "transformers>=4.40.1",
    "trl>=0.12.0",
    "peft",
    "accelerate",
    "bitsandbytes",
    "wandb",
    "pyyaml",
    "unsloth @ git+https://github.com/unslothai/unsloth.git",
    "xformers",
    "packaging",
    "ninja",
    "datasets",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["scripts", "configs", "training"]

[tool.uv]
dev-dependencies = [
    "ipykernel",
    "jupyter",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"